packages/czlai/czlai-0.1.0a6.tar.gz/czlai-0.1.0a6/tests/api_resources/users/test_industry.py
# File generated from our OpenAPI spec by Stainless. See CONTRIBUTING.md for details.

from __future__ import annotations

import os
from typing import Any, cast

import pytest

from czlai import Czlai, AsyncCzlai

base_url = os.environ.get("TEST_API_BASE_URL", "http://127.0.0.1:4010")


class TestIndustry:
    parametrize = pytest.mark.parametrize("client", [False, True], indirect=True, ids=["loose", "strict"])

    @parametrize
    def test_method_retrieve(self, client: Czlai) -> None:
        industry = client.users.industry.retrieve()
        assert industry is None

    @parametrize
    def test_raw_response_retrieve(self, client: Czlai) -> None:
        response = client.users.industry.with_raw_response.retrieve()

        assert response.is_closed is True
        assert response.http_request.headers.get("X-Stainless-Lang") == "python"
        industry = response.parse()
        assert industry is None

    @parametrize
    def test_streaming_response_retrieve(self, client: Czlai) -> None:
        with client.users.industry.with_streaming_response.retrieve() as response:
            assert not response.is_closed
            assert response.http_request.headers.get("X-Stainless-Lang") == "python"

            industry = response.parse()
            assert industry is None

        assert cast(Any, response.is_closed) is True


class TestAsyncIndustry:
    parametrize = pytest.mark.parametrize("async_client", [False, True], indirect=True, ids=["loose", "strict"])

    @parametrize
    async def test_method_retrieve(self, async_client: AsyncCzlai) -> None:
        industry = await async_client.users.industry.retrieve()
        assert industry is None

    @parametrize
    async def test_raw_response_retrieve(self, async_client: AsyncCzlai) -> None:
        response = await async_client.users.industry.with_raw_response.retrieve()

        assert response.is_closed is True
        assert response.http_request.headers.get("X-Stainless-Lang") == "python"
        industry = await response.parse()
        assert industry is None

    @parametrize
    async def test_streaming_response_retrieve(self, async_client: AsyncCzlai) -> None:
        async with async_client.users.industry.with_streaming_response.retrieve() as response:
            assert not response.is_closed
            assert response.http_request.headers.get("X-Stainless-Lang") == "python"

            industry = await response.parse()
            assert industry is None

        assert cast(Any, response.is_closed) is True
